# src/main/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flowrider API
  description: |
    REST API for Flowrider - AI Session Orchestration Platform

    Flowrider is an Electron app for orchestrating multiple AI terminal sessions using a 3D icosahedron interface (20 faces = 20 sessions).

    ## Authentication
    All API requests require authentication using an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per API key.

    ## Error Handling
    The API uses standard HTTP status codes and returns error details in JSON format.
  version: 1.0.0
  contact:
    name: Josephas LLC
    url: https://github.com/josephas-llc/flowrider
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3737/api
    description: Local development server
  - url: https://api.flowrider.local/api
    description: Production server (when deployed)
...
security:
  - ApiKeyAuth: []
paths:
...
  /keys:
    get:
      tags:
        - API Keys
      summary: List all API keys
      description: Retrieve a list of all API keys
      operationId: listApiKeys
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - API Keys
      summary: Create a new API key
      description: Generate a new API key
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Descriptive name for the API key
                  example: "CI/CD Pipeline"
                scopes:
                  type: array
                  description: Permission scopes
                  items:
                    type: string
                    enum:
                      - sessions:read
                      - sessions:write
                      - projects:read
                      - projects:write
                      - webhooks:read
                      - webhooks:write
                      - keys:read
                      - keys:write
                  example: ["sessions:read", "sessions:write", "projects:read"]
                expiresAt:
                  type: string
                  format: date-time
                  description: Optional expiration date
                  example: "2026-12-31T23:59:59Z"
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    allOf:
                      - $ref: '#/components/schemas/ApiKey'
                      - type: object
                        properties:
                          key:
                            type: string
                            description: The actual API key (only shown once)
                            example: "fwr_live_abc123def456ghi789"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /keys/{id}:
    delete:
      tags:
        - API Keys
      summary: Delete API key
      description: Revoke an API key
      operationId: deleteApiKey
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '200':
          description: API key deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "API key revoked successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ========================================
  # HEALTH
  # ========================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  parameters:
...
    ApiKeyId:
      name: id
      in: path
      required: true
      description: API Key ID
      schema:
        type: string
        example: "key_789ghi"
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          example: "key_789ghi"
        name:
          type: string
          example: "CI/CD Pipeline"
        prefix:
          type: string
          description: First few characters of the key
          example: "fwr_live_abc"
        scopes:
          type: array
          items:
            type: string
          example: ["sessions:read", "sessions:write"]
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-07-06T15:30:00Z"
        expiresAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-12-31T23:59:59Z"
        createdAt:
          type: string
          format: date-time
          example: "2026-07-01T10:00:00Z"
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid request parameters"
        code:
          type: string
          description: Error code
          example: "INVALID_INPUT"
        details:
          type: object
          nullable: true
          description: Additional error details
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid request parameters"
            code: "INVALID_INPUT"
    UnauthorizedError:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid or missing API key"
            code: "UNAUTHORIZED"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "An internal server error occurred"
            code: "INTERNAL_ERROR"